Precompile the template for timestamps
There's slightly better performance and it's less annoying to use.
One eye opening thing is comparing the code generated with and without
the *variable* option. Without it uses a with keyword, which is one of
the slowest things you can do in a JS engine as it's very hard to
optimize.