WIP - SRT format parser
Passes initial tests. Needs a bunch of refactoring and tests involving
longer sequences of SRT.

Uses jQuery Deferred for the API. I want to experiment with making async
APIs using only Deferreds to understand the advantages and
disadvantages. If this works I may try out Q.

The fromSrt function accepts two parameters, one an array of strings
containing the source data, two the collection to add the Subtitles too.

It returns a jQuery promise and runs async. It's not clear if this will
take a long time for the average SRT file. If it does, the task may have
to add progress calls to yield after a certain number of SRTs have been
created.

The done callback returns the number of lines parsed and the collection.
The fail callback returns the error.