Issue #1 Fix JS double bug
Because JS only has doubles and not true ints, the remainder operation
also passes through the mantissa. Run a Math.floor function on the
result to make sure it's rounded down to the nearest int.

Add a unit test to catch this. All the previous units used exact
multiples of all the units and so didn't trigger the rounding
problem.