Rough and ready proof of concept player
This is a working player, but it's a proof of concept.

It has a number of drawbacks. It can't handle overlapping subtitles. I'm
not sure it's in the spec I have to check. It can't pause and resume.
It plays even when subtitles aren't loaded. Stepping doesn't
work yet.

I'm planning on moving the timer code into it's own event based trigger.
Same thing for the player. Marionette has an event based system
built in.

The button controls are also flakey.

Yuhri would also like me to stream the video at the same time.
This will probably require live transcoding.