Basic wiring for new Subtitle Player Architecture
Using a Layout which is contained in a Region.  The Layout has 3
Sub Regions, which each contain a view, one for each element of the
Player.

I'm using an EventAggregator to communicate between
the Views so they are loosely coupled.

The eventual plan is to send Event to app.vent for start and stop
events so I can synchronize with the current subtitle metadata
in the scroll window.